requirements_dev.txt
# Testing tools and code formatters
pytest
coverage
pytest-cov
black
pyyaml

# Project dependencies
networkx
lark-parser

